Improve Ruby spec compliance: method hooks, Kernel refactor, builtins
- Move method implementations from Object to Kernel module to match CRuby
- Implement method_missing fallback with argument passing
- Add singleton_method_added/removed/undefined hooks
- Consolidate method_added/removed/undefined callbacks into Executor helpers
- Add Method#arity, NoMethodError#receiver, Module#class_variables
- Fix != dispatch, protected method check, frozen singleton, constant lookup
- Reject singleton class creation for Integer, Float, Symbol, Range, Complex
- Define BasicObject as constant on itself, improve spec compliance
- Add comprehensive tests for method_call and other builtins
