+++ monoruby/src/builtins/kernel.rs
@@ -115,6 +115,74 @@ pub(super) fn init(globals: &mut Globals) -> Module {
     //globals.define_builtin_inline_func(kernel_class, "____max", max, Box::new(inline_max), 2);
     //globals.define_builtin_inline_func(kernel_class, "____min", min, Box::new(inline_min), 2);
 
+    // Kernel methods (matching CRuby: these are defined on Kernel, not Object)
+    globals.define_builtin_func(kernel_class, "class", class, 0);
+    globals.define_builtin_func(kernel_class, "hash", hash, 0);
+    globals.define_builtin_func(kernel_class, "eql?", eql_, 1);
+    globals.define_builtin_funcs(kernel_class, "dup", &["clone"], dup, 0);
+    globals.define_private_builtin_func(kernel_class, "initialize_copy", initialize_copy, 1);
+    globals.define_private_builtin_func(kernel_class, "initialize_clone", initialize_clone, 1);
+    globals.define_private_builtin_func(kernel_class, "initialize_dup", initialize_clone, 1);
+    globals.define_builtin_funcs_rest(kernel_class, "enum_for", &["to_enum"], to_enum);
+    globals.define_builtin_func_rest(kernel_class, "extend", extend);
+    globals.define_builtin_func(kernel_class, "kind_of?", is_a, 1);
+    globals.define_builtin_inline_func(
+        kernel_class,
+        "object_id",
+        super::object::object_id,
+        Box::new(super::object::object_object_id),
+        0,
+    );
+    globals.define_builtin_inline_func_with(
+        kernel_class,
+        "respond_to?",
+        respond_to,
+        Box::new(object_respond_to),
+        1,
+        2,
+        false,
+    );
+    globals.define_builtin_func(kernel_class, "singleton_class", singleton_class, 0);
+    globals.define_builtin_func(kernel_class, "to_s", to_s, 0);
+    globals.define_builtin_func(kernel_class, "inspect", inspect, 0);
+    globals.define_builtin_func(kernel_class, "instance_of?", instance_of, 1);
+    globals.define_builtin_func(kernel_class, "instance_variable_defined?", iv_defined, 1);
+    globals.define_builtin_func(kernel_class, "instance_variable_set", iv_set, 2);
+    globals.define_builtin_func(kernel_class, "instance_variable_get", iv_get, 1);
+    globals.define_builtin_func(kernel_class, "instance_variables", iv, 0);
+    globals.define_builtin_func(kernel_class, "remove_instance_variable", iv_remove, 1);
+    globals.define_builtin_func(kernel_class, "is_a?", is_a, 1);
+    globals.define_builtin_inline_funcs_with_kw(
+        kernel_class,
+        "send",
+        &["__send__", "public_send"],
+        crate::builtins::send,
+        Box::new(crate::builtins::object_send),
+        0,
+        0,
+        true,
+        &[],
+        true,
+    );
+    globals.define_builtin_func(kernel_class, "method", method, 1);
+    globals.define_builtin_func_with(
+        kernel_class,
+        "define_singleton_method",
+        define_singleton_method,
+        1,
+        2,
+        false,
+    );
+    globals.define_builtin_func_with(kernel_class, "methods", methods, 0, 1, false);
+    globals.define_builtin_func_with(
+        kernel_class,
+        "singleton_methods",
+        singleton_methods,
+        0,
+        1,
+        false,
+    );
+
     klass
 }
 
@@ -769,7 +837,7 @@ fn eval(vm: &mut Executor, globals: &mut Globals, lfp: Lfp, _: BytecodePtr) -> R
         vm.invoke_binding(globals, binding.binding().unwrap())
     } else {
         let caller_cfp = cfp.prev().unwrap();
-        let fid = globals.compile_script_eval(expr, "(eval)", caller_cfp)?;
+        let fid = globals.compile_script_eval(expr, "(eval)", caller_cfp, None)?;
         let proc = ProcData::new(caller_cfp.lfp(), fid);
         vm.invoke_block(globals, &proc, &[])
     }
@@ -1428,6 +1496,540 @@ fn inline_min(
     true
 }*/
 
+///
+/// Object#send
+///
+/// - send(name, *args) -> object
+/// - send(name, *args) { .... } -> object
+///
+/// [https://docs.ruby-lang.org/ja/latest/method/Object/i/send.html]
+#[monoruby_builtin]
+pub(crate) fn send(
+    vm: &mut Executor,
+    globals: &mut Globals,
+    lfp: Lfp,
+    _: BytecodePtr,
+) -> Result<Value> {
+    let ary = lfp.arg(0).as_array();
+    if ary.len() < 1 {
+        return Err(MonorubyErr::wrong_number_of_arg_min(ary.len(), 1));
+    }
+    let method = ary[0].expect_symbol_or_string(globals)?;
+    vm.invoke_method_inner(
+        globals,
+        method,
+        lfp.self_val(),
+        &ary[1..],
+        lfp.block(),
+        if let Some(kw) = lfp.try_arg(1)
+            && let Some(kw) = kw.try_hash_ty()
+            && !kw.is_empty()
+        {
+            Some(kw)
+        } else {
+            None
+        },
+    )
+}
+
+pub fn object_send(
+    state: &mut AbstractState,
+    ir: &mut AsmIr,
+    _: &JitContext,
+    store: &Store,
+    callid: CallSiteId,
+    _: ClassId,
+) -> bool {
+    let callsite = &store[callid];
+    let no_splat = !callsite.object_send_single_splat();
+    if !callsite.is_simple() && no_splat {
+        return false;
+    }
+
+    state.write_back_recv_and_callargs(ir, callsite);
+    state.writeback_acc(ir);
+    let using_xmm = state.get_using_xmm();
+    let error = ir.new_error(state);
+    let callid = callsite.id;
+    ir.inline(move |r#gen, store, labels| {
+        let error = &labels[error];
+        r#gen.object_send_inline(callid, store, using_xmm, &error, no_splat);
+    });
+    state.def_reg2acc(ir, GP::Rax, callsite.dst);
+    true
+}
+
+///
+/// ### Kernel#extend
+/// - extend(*modules) -> self
+///
+/// [https://docs.ruby-lang.org/ja/latest/method/Object/i/extend.html]
+#[monoruby_builtin]
+fn extend(vm: &mut Executor, globals: &mut Globals, lfp: Lfp, _: BytecodePtr) -> Result<Value> {
+    let args = lfp.arg(0).as_array();
+    if args.len() == 0 {
+        return Err(MonorubyErr::wrong_number_of_arg_min(0, 1));
+    }
+    let self_val = lfp.self_val();
+    for v in args.iter().cloned().rev() {
+        vm.invoke_method_inner(globals, IdentId::EXTEND_OBJECT, v, &[self_val], None, None)?;
+        vm.invoke_method_inner(globals, IdentId::EXTENDED, v, &[self_val], None, None)?;
+    }
+    Ok(lfp.self_val())
+}
+
+///
+/// ### Kernel#is_a?
+///
+/// - is_a?(mod) -> bool
+/// - kind_of?(mod) -> bool
+///
+/// [https://docs.ruby-lang.org/ja/latest/method/Object/i/is_a=3f.html]
+#[monoruby_builtin]
+fn is_a(_vm: &mut Executor, globals: &mut Globals, lfp: Lfp, _: BytecodePtr) -> Result<Value> {
+    let class = lfp.arg(0).expect_class_or_module(&globals.store)?.id();
+    Ok(Value::bool(
+        lfp.self_val().is_kind_of(&globals.store, class),
+    ))
+}
+
+///
+/// ### Kernel#enum_for
+///
+/// - to_enum(method = :each, *args) -> Enumerator
+/// - enum_for(method = :each, *args) -> Enumerator
+/// - [NOT SUPPORTED] to_enum(method = :each,  *args) {|*args| ... } -> Enumerator
+/// - [NOT SUPPORTED] enum_for(method = :each, *args) {|*args| ... } -> Enumerator
+///
+/// [https://docs.ruby-lang.org/ja/latest/method/Object/i/enum_for.html]
+#[monoruby_builtin]
+fn to_enum(vm: &mut Executor, globals: &mut Globals, lfp: Lfp, pc: BytecodePtr) -> Result<Value> {
+    lfp.expect_no_block()?;
+    let args = lfp.arg(0).as_array();
+    let (method, args) = if args.is_empty() {
+        (IdentId::EACH, vec![])
+    } else {
+        (
+            args[0].expect_symbol_or_string(globals)?,
+            args[1..].to_vec(),
+        )
+    };
+    vm.generate_enumerator(method, lfp.self_val(), args, pc)
+}
+
+///
+/// ### Kernel#dup
+///
+/// - dup -> object
+///
+/// [https://docs.ruby-lang.org/ja/latest/method/Object/i/clone.html]
+#[monoruby_builtin]
+fn dup(_vm: &mut Executor, _globals: &mut Globals, lfp: Lfp, _: BytecodePtr) -> Result<Value> {
+    Ok(lfp.self_val().dup())
+}
+
+/// ### Kernel#define_singleton_method
+///
+/// - define_singleton_method(name, method) -> Symbol
+/// - define_singleton_method(name) { ... } -> Symbol
+///
+/// [https://docs.ruby-lang.org/ja/latest/method/Object/i/define_singleton_method.html]
+#[monoruby_builtin]
+fn define_singleton_method(
+    vm: &mut Executor,
+    globals: &mut Globals,
+    lfp: Lfp,
+    pc: BytecodePtr,
+) -> Result<Value> {
+    let self_val = lfp.self_val();
+    let class_id = globals.store.get_singleton(self_val)?.id();
+    let name = lfp.arg(0).expect_symbol_or_string(globals)?;
+    let func_id = if let Some(method) = lfp.try_arg(1) {
+        if let Some(proc) = method.is_proc() {
+            globals.define_proc_method(proc)
+        } else if let Some(method) = method.is_method() {
+            method.func_id()
+        } else if let Some(method) = method.is_umethod() {
+            method.func_id()
+        } else {
+            return Err(MonorubyErr::wrong_argument_type(
+                globals,
+                method,
+                "Proc/Method/UnboundMethod",
+            ));
+        }
+    } else if let Some(bh) = lfp.block() {
+        let proc = vm.generate_proc(bh, pc)?;
+        globals.define_proc_method(proc)
+    } else {
+        return Err(MonorubyErr::wrong_number_of_arg(2, 1));
+    };
+    globals.add_public_method(class_id, name, func_id);
+    Ok(Value::symbol(name))
+}
+
+/// ### Kernel#initialize_copy
+///
+/// - initialize_copy(obj) -> object
+///
+/// [https://docs.ruby-lang.org/ja/latest/method/Object/i/initialize_copy.html]
+#[monoruby_builtin]
+fn initialize_copy(
+    _: &mut Executor,
+    globals: &mut Globals,
+    lfp: Lfp,
+    _: BytecodePtr,
+) -> Result<Value> {
+    let orig = lfp.arg(0);
+    let self_val = lfp.self_val();
+    if orig.id() == self_val.id() {
+        return Ok(self_val);
+    }
+    if self_val.real_class(globals).id() != orig.real_class(globals).id()
+        || self_val.ty() != orig.ty()
+    {
+        return Err(MonorubyErr::typeerr(format!(
+            "initialize_copy should take same class object self:{} original:{}",
+            self_val.class().get_name(globals),
+            orig.class().get_name(globals)
+        )));
+    }
+    Ok(self_val)
+}
+
+/// ### Kernel#initialize_clone
+///
+/// - initialize_clone(obj) -> object
+///
+/// [https://docs.ruby-lang.org/ja/latest/method/Object/i/initialize_clone.html]
+#[monoruby_builtin]
+fn initialize_clone(
+    vm: &mut Executor,
+    globals: &mut Globals,
+    lfp: Lfp,
+    _: BytecodePtr,
+) -> Result<Value> {
+    let orig = lfp.arg(0);
+    let self_val = lfp.self_val();
+    vm.invoke_method_inner(
+        globals,
+        IdentId::get_id("initialize_copy"),
+        self_val,
+        &[orig],
+        None,
+        None,
+    )
+}
+
+///
+/// ### Kernel#to_s
+///
+/// - to_s -> String
+///
+/// [https://docs.ruby-lang.org/ja/latest/method/Object/i/to_s.html]
+#[monoruby_builtin]
+fn to_s(_vm: &mut Executor, globals: &mut Globals, lfp: Lfp, _: BytecodePtr) -> Result<Value> {
+    let s = lfp.self_val().to_s(&globals.store);
+    Ok(Value::string(s))
+}
+
+///
+/// ### Kernel#respond_to?
+///
+/// - respond_to?(name, include_all = false) -> bool
+///
+/// [https://docs.ruby-lang.org/ja/latest/method/Object/i/respond_to=3f.html]
+#[monoruby_builtin]
+fn respond_to(
+    _vm: &mut Executor,
+    globals: &mut Globals,
+    lfp: Lfp,
+    _: BytecodePtr,
+) -> Result<Value> {
+    let name = lfp.arg(0).expect_symbol_or_string(globals)?;
+    let include_all = if let Some(arg1) = lfp.try_arg(1) {
+        arg1.as_bool()
+    } else {
+        false
+    };
+    Ok(Value::bool(if include_all {
+        globals.check_method(lfp.self_val(), name).is_some()
+    } else {
+        globals.check_public_method(lfp.self_val(), name).is_some()
+    }))
+}
+
+fn object_respond_to(
+    state: &mut AbstractState,
+    _: &mut AsmIr,
+    ctx: &JitContext,
+    store: &Store,
+    callid: CallSiteId,
+    recv_class: ClassId,
+) -> bool {
+    let callsite = &store[callid];
+    if !callsite.is_simple() {
+        return false;
+    }
+    let CallSiteInfo {
+        dst, args, pos_num, ..
+    } = *callsite;
+    let dst = if let Some(dst) = dst {
+        dst
+    } else {
+        return false;
+    };
+    let include_all = if pos_num != 1 {
+        if state.is_truthy(args + 1usize) {
+            true
+        } else if state.is_falsy(args + 1usize) {
+            false
+        } else {
+            return false;
+        }
+    } else {
+        false
+    };
+    let method_name = if let Some(name) = state.is_symbol_literal(args) {
+        name
+    } else {
+        return false;
+    };
+    let b = if let Some(entry) =
+        store.check_method_for_class_with_version(recv_class, method_name, ctx.class_version())
+    {
+        include_all || entry.is_public()
+    } else {
+        false
+    };
+    state.def_C(dst, Value::bool(b));
+    true
+}
+
+///
+/// ### Kernel#inspect
+///
+/// - inspect -> String
+///
+/// [https://docs.ruby-lang.org/ja/latest/method/Object/i/inspect.html]
+#[monoruby_builtin]
+fn inspect(_vm: &mut Executor, globals: &mut Globals, lfp: Lfp, _: BytecodePtr) -> Result<Value> {
+    let s = lfp.self_val().inspect(&globals.store);
+    Ok(Value::string(s))
+}
+
+///
+/// ### Kernel#class
+///
+/// - class -> Class
+///
+/// [https://docs.ruby-lang.org/ja/latest/method/Object/i/class.html]
+#[monoruby_builtin]
+fn class(_vm: &mut Executor, globals: &mut Globals, lfp: Lfp, _: BytecodePtr) -> Result<Value> {
+    Ok(lfp.self_val().real_class(&globals.store).as_val())
+}
+
+///
+/// ### Kernel#hash
+///
+/// - hash -> Integer
+///
+/// [https://docs.ruby-lang.org/ja/latest/method/Object/i/hash.html]
+#[monoruby_builtin]
+fn hash(_: &mut Executor, _: &mut Globals, lfp: Lfp, _: BytecodePtr) -> Result<Value> {
+    Ok(Value::integer(lfp.self_val().id() as _))
+}
+
+///
+/// ### Kernel#eql?
+///
+/// - eql?(other) -> bool
+///
+/// [https://docs.ruby-lang.org/ja/latest/method/Object/i/eql=3f.html]
+#[monoruby_builtin]
+fn eql_(_: &mut Executor, _: &mut Globals, lfp: Lfp, _: BytecodePtr) -> Result<Value> {
+    Ok(Value::bool(lfp.self_val().id() == lfp.arg(0).id()))
+}
+
+///
+/// ### Kernel#instance_of?
+///
+/// - instance_of?(klass) -> bool
+///
+/// [https://docs.ruby-lang.org/ja/latest/method/Object/i/instance_of=3f.html]
+#[monoruby_builtin]
+fn instance_of(
+    _vm: &mut Executor,
+    globals: &mut Globals,
+    lfp: Lfp,
+    _: BytecodePtr,
+) -> Result<Value> {
+    let b = lfp.self_val().real_class(&globals.store).id()
+        == lfp.arg(0).expect_class_or_module(&globals.store)?.id();
+    Ok(Value::bool(b))
+}
+
+///
+/// ### Kernel#method
+///
+/// - method(name) -> Method
+///
+/// [https://docs.ruby-lang.org/ja/latest/method/Object/i/method.html]
+#[monoruby_builtin]
+fn method(_vm: &mut Executor, globals: &mut Globals, lfp: Lfp, _: BytecodePtr) -> Result<Value> {
+    let receiver = lfp.self_val();
+    let method_name = lfp.arg(0).expect_symbol_or_string(globals)?;
+    let (func_id, _, owner) = globals.find_method_for_object(receiver, method_name)?;
+    Ok(Value::new_method(receiver, func_id, owner))
+}
+
+///
+/// ### Kernel#singleton_class
+///
+/// - singleton_class -> Class
+///
+/// [https://docs.ruby-lang.org/ja/latest/method/Object/i/singleton_class.html]
+#[monoruby_builtin]
+fn singleton_class(
+    _vm: &mut Executor,
+    globals: &mut Globals,
+    lfp: Lfp,
+    _: BytecodePtr,
+) -> Result<Value> {
+    lfp.self_val().get_singleton(&mut globals.store)
+}
+
+///
+/// ### Kernel#methods
+///
+/// - methods(include_inherited = true) -> [Symbol]
+///
+/// [https://docs.ruby-lang.org/ja/latest/method/Object/i/methods.html]
+#[monoruby_builtin]
+fn methods(_vm: &mut Executor, globals: &mut Globals, lfp: Lfp, _: BytecodePtr) -> Result<Value> {
+    let inherited_too = lfp.try_arg(0).is_none() || lfp.arg(0).as_bool();
+    Ok(Value::array_from_vec(if !inherited_too {
+        let class_id = match globals.store.has_singleton(lfp.self_val()) {
+            Some(module) => module.id(),
+            None => return Ok(Value::array_empty()),
+        };
+        globals.store.get_method_names(class_id)
+    } else {
+        let class_id = lfp.self_val().class();
+        globals.store.get_method_names_inherit(class_id, true)
+    }))
+}
+
+///
+/// ### Kernel#singleton_methods
+///
+/// - singleton_methods(inherited_too = true) -> [Symbol]
+///
+/// [https://docs.ruby-lang.org/ja/latest/method/Object/i/singleton_methods.html]
+#[monoruby_builtin]
+fn singleton_methods(
+    _vm: &mut Executor,
+    globals: &mut Globals,
+    lfp: Lfp,
+    _: BytecodePtr,
+) -> Result<Value> {
+    let class_id = match globals.store.has_singleton(lfp.self_val()) {
+        Some(module) => module.id(),
+        None => return Ok(Value::array_empty()),
+    };
+    let inherited_too = lfp.try_arg(0).is_none() || lfp.arg(0).as_bool();
+    Ok(Value::array_from_vec(if !inherited_too {
+        globals.store.get_method_names(class_id)
+    } else {
+        globals.store.get_method_names_inherit(class_id, true)
+    }))
+}
+
+///
+/// ### Kernel#instance_variable_defined?
+///
+/// - instance_variable_defined?(var) -> bool
+///
+/// [https://docs.ruby-lang.org/ja/latest/method/Object/i/instance_variable_defined=3f.html]
+#[monoruby_builtin]
+fn iv_defined(
+    _vm: &mut Executor,
+    globals: &mut Globals,
+    lfp: Lfp,
+    _: BytecodePtr,
+) -> Result<Value> {
+    let id = match lfp.arg(0).unpack() {
+        RV::Symbol(sym) => sym,
+        RV::String(s) => IdentId::get_id(s.check_utf8()?),
+        _ => return Err(MonorubyErr::is_not_symbol_nor_string(globals, lfp.arg(0))),
+    };
+    let b = globals.store.get_ivar(lfp.self_val(), id).is_some();
+    Ok(Value::bool(b))
+}
+
+///
+/// ### Kernel#instance_variable_set
+///
+/// - instance_variable_set(var, value) -> Object
+///
+/// [https://docs.ruby-lang.org/ja/latest/method/Object/i/instance_variable_set.html]
+#[monoruby_builtin]
+fn iv_set(_vm: &mut Executor, globals: &mut Globals, lfp: Lfp, _: BytecodePtr) -> Result<Value> {
+    let id = lfp.arg(0).expect_symbol_or_string(globals)?;
+    let val = lfp.arg(1);
+    globals.store.set_ivar(lfp.self_val(), id, val)?;
+    Ok(val)
+}
+
+///
+/// ### Kernel#instance_variable_get
+///
+/// - instance_variable_get(var) -> Object | nil
+///
+/// [https://docs.ruby-lang.org/ja/latest/method/Object/i/instance_variable_get.html]
+#[monoruby_builtin]
+fn iv_get(_vm: &mut Executor, globals: &mut Globals, lfp: Lfp, _: BytecodePtr) -> Result<Value> {
+    let id = lfp.arg(0).expect_symbol_or_string(globals)?;
+    let v = globals
+        .store
+        .get_ivar(lfp.self_val(), id)
+        .unwrap_or_default();
+    Ok(v)
+}
+
+///
+/// ### Kernel#instance_variables
+///
+/// - instance_variables -> [Symbol]
+///
+/// [https://docs.ruby-lang.org/ja/latest/method/Object/i/instance_variables.html]
+#[monoruby_builtin]
+fn iv(_vm: &mut Executor, globals: &mut Globals, lfp: Lfp, _: BytecodePtr) -> Result<Value> {
+    let iter = globals
+        .store
+        .get_ivars(lfp.self_val())
+        .into_iter()
+        .map(|(id, _)| Value::symbol(id));
+    Ok(Value::array_from_iter(iter))
+}
+
+///
+/// ### Kernel#remove_instance_variable
+///
+/// - remove_instance_variable(name) -> Object
+///
+/// [https://docs.ruby-lang.org/ja/latest/method/Object/i/remove_instance_variable.html]
+#[monoruby_builtin]
+fn iv_remove(_vm: &mut Executor, globals: &mut Globals, lfp: Lfp, _: BytecodePtr) -> Result<Value> {
+    let id = lfp.arg(0).expect_symbol_or_string(globals)?;
+    match globals.store.remove_ivar(lfp.self_val(), id) {
+        Some(val) => Ok(val),
+        None => Err(MonorubyErr::nameerr(format!(
+            "instance variable {id} not defined"
+        ))),
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use crate::tests::*;
@@ -1751,6 +2353,736 @@ mod tests {
         end
         p PFoo.new
         "##,
+        );
+    }
+
+    #[test]
+    fn test_builtin() {
+        run_test2(":sym.class.to_s");
+        run_test2("5.class.to_s");
+        run_test2("5.7.class.to_s");
+        run_test2("'windows'.class.to_s");
+        run_test2("puts 100");
+        run_test2("print '100'");
+        run_test2("p");
+        run_test2("p 1");
+        run_test2("p 1,2");
+        run_test2("p 1,2,3");
+        run_test2("nil.respond_to?(:foo)");
+        run_test2("nil.inspect");
+        run_test2("Time.singleton_class.to_s");
+        run_test2(r#"File.write("/tmp/foo", "woo")"#);
+    }
+
+    #[test]
+    fn object_id() {
+        run_test_with_prelude(
+            r##"
+            [id == a.object_id, a.__id__ == a.object_id]
+            "##,
+            r##"
+            a = [1,2,3]
+            id = a.object_id
+            "##,
+        );
+        run_test_with_prelude(
+            r##"
+            [id == a.object_id, a.__id__ == a.object_id]
+            "##,
+            r##"
+            a = 1356
+            id = a.object_id
+            "##,
+        );
+        run_test_with_prelude(
+            r##"
+            [id == a.object_id, a.__id__ == a.object_id]
+            "##,
+            r##"
+            a = -49.52
+            id = a.object_id
+            "##,
+        );
+    }
+
+    #[test]
+    fn remove_instance_variable() {
+        run_test2(
+            r#"
+            class Foo
+              def initialize
+                @a = 1
+                @b = 2
+              end
+            end
+            f = Foo.new
+            f.remove_instance_variable(:@a)
+            "#,
+        );
+        run_test2(
+            r#"
+            class Foo
+              def initialize
+                @a = 1
+                @b = 2
+              end
+            end
+            f = Foo.new
+            f.remove_instance_variable(:@a)
+            f.instance_variables
+            "#,
+        );
+        run_test2(
+            r#"
+            class Foo
+              def initialize
+                @a = 1
+                @b = 2
+              end
+            end
+            f = Foo.new
+            f.remove_instance_variable("@a")
+            f.instance_variables
+            "#,
+        );
+        run_test2(
+            r#"
+            module M
+              @x = 42
+              remove_instance_variable(:@x)
+            end
+            "#,
+        );
+        run_test_error(
+            r#"
+            obj = Object.new
+            obj.remove_instance_variable(:@nonexistent)
+            "#,
+        );
+    }
+
+    #[test]
+    fn extend() {
+        run_test_with_prelude(
+            r#"
+        res = []
+        obj = Object.new
+        obj.extend Foo, Bar
+        res << obj.a #=> "ok Foo"
+        res << obj.b #=> "ok Bar"
+        res << Klass.new.a #=> "ok Foo"
+        res << Klass.b     #=> "ok Bar"
+        res
+        "#,
+            r#"
+        module Foo
+          def a
+            'ok Foo'
+          end
+        end
+
+        module Bar
+          def b
+            'ok Bar'
+          end
+        end
+
+        class Klass
+          include Foo
+          extend Bar
+        end
+        "#,
+        );
+    }
+
+    #[test]
+    fn dup() {
+        run_test("1.dup");
+        run_test("1.5.dup");
+        run_test("'Ruby'.dup");
+        run_test(":Ruby.dup");
+        run_test("[1,2,3].dup");
+        run_test("{a:1,b:2}.dup");
+        run_test("(1..3).dup");
+    }
+
+    #[test]
+    fn block_given_kernel() {
+        run_test_with_prelude(
+            r##"
+            [check, check {}]
+        "##,
+            r##"
+        def check
+            if block_given?
+                "Block is given."
+            else
+                "Block isn't given."
+            end
+        end
+        "##,
+        );
+    }
+
+    #[test]
+    fn test_object() {
+        run_test2(r#"a=Object.new; a.instance_variable_set("@i", 42)"#);
+        run_test2(r#"a=Object.new; a.instance_variable_get(:@i)"#);
+        run_test2(
+            r#"a=Object.new; a.instance_variable_set("@i", 42); a.instance_variable_defined?(:@i)"#,
+        );
+        run_test2(
+            r#"a=Object.new; a.instance_variable_set("@i", 42); a.instance_variable_defined?(:@j)"#,
+        );
+        run_test2(
+            r#"a=Object.new; a.instance_variable_set("@i", 42); a.instance_variable_get(:@i)"#,
+        );
+        // TDDO: Object#instace_variables must return Array of Symbols,
+        // ordered by the time of definition of the instance variables.
+        run_test2(
+            r#"
+            a=Object.new;
+            a.instance_variable_set("@b", 1);
+            a.instance_variable_set("@e", 4);
+            a.instance_variable_set("@c", 2);
+            a.instance_variable_set("@j", 9);
+            a.instance_variable_set("@d", 3);
+            a.instance_variable_set("@a", 0);
+            a.instance_variable_set("@g", 6);
+            a.instance_variable_set("@l", 11);
+            a.instance_variable_set("@h", 7);
+            a.instance_variable_set("@i", 8);
+            a.instance_variable_set("@f", 5);
+            a.instance_variable_set("@k", 10);
+            a.instance_variables.sort
+            "#,
+        );
+    }
+
+    #[test]
+    fn respond_to() {
+        run_test_with_prelude(
+            r#"
+        list = [F.new, D.new]
+        res = []
+        list.each{|it| res << it.hello if it.respond_to?(:hello)}
+        list.each{|it| res << it.hello if it.respond_to?(:hello, false)}
+        list.each{|it| it.instance_eval("res << hello if it.respond_to?(:hello, true)")}
+        res
+        "#,
+            r#"
+        class F
+          def hello
+            "Bonjour"
+          end
+        end
+
+        class D
+          private
+          def hello
+            "Guten Tag"
+          end
+        end
+        "#,
+        );
+    }
+
+    #[test]
+    fn kernel_integer() {
+        run_test2(r#"Integer(-2435)"#);
+        run_test2(r#"Integer("2435")"#);
+        run_test_error(r#"Integer("2435.78")"#);
+        run_test_error(r#"Integer([4])"#);
+        run_test2(r#"Integer(-2435766756886769978978435)"#);
+        run_test2(r#"Integer(2435.4556787)"#);
+    }
+
+    #[test]
+    fn object_instance_of() {
+        run_test2(r#"5.instance_of?(Integer)"#);
+        run_test2(r#"5.instance_of?(Float)"#);
+        run_test2(r#":ruby.instance_of?(Symbol)"#);
+        run_test2(r#":ruby.instance_of?(Object)"#);
+        run_test2(
+            r#"
+        class C < Object
+        end
+        class S < C
+        end
+
+        obj = S.new
+        [obj.instance_of?(S), obj.instance_of?(C)]
+        "#,
+        );
+        run_test_error(r#"5.instance_of?(7)"#);
+    }
+
+    #[test]
+    fn object_isa() {
+        run_test("4.is_a? Integer");
+        run_test("4.5.is_a? Integer");
+        run_test("'Ruby'.is_a? Integer");
+        run_test("4.5.is_a? Float");
+        run_test("'Ruby'.is_a? Float");
+        run_test(
+            r#"
+        class C
+        end
+        C.new.is_a? C"#,
+        );
+        run_test(
+            r#"
+        class S
+        end
+        class C < S
+        end
+        c = C.new
+        [c.is_a?(S), c.is_a?(C)]"#,
+        );
+    }
+
+    #[test]
+    fn object_nil() {
+        run_test("4.nil?");
+        run_test("4.5.nil?");
+        run_test("nil.nil?");
+        run_test("true.nil?");
+        run_test("false.nil?");
+        run_test("[].nil?");
+    }
+
+    #[test]
+    fn kernel_system() {
+        run_test(r#"system "ls""#);
+        run_test(r#"system "jkjkjk""#);
+        run_test(r#"system "*""#);
+        run_test(r#"`pwd`"#);
+        run_test(r#"`*`"#);
+        run_test_error(r#"``"#);
+    }
+
+    #[test]
+    fn to_enum() {
+        run_test(
+            r#"
+        e = [1,2,3,4,5].to_enum
+        e.next
+        "#,
+        );
+    }
+
+    #[test]
+    fn send() {
+        run_test(
+            r#"
+        "ruby".send(:sub, /./, "R")
+        "#,
+        );
+        run_test(
+            r#"
+        class Foo
+            def foo() "foo" end
+            def bar() "bar" end
+            def baz() "baz" end
+        end
+        methods = {1 => :foo, 2 => :bar, 3 => :baz}
+        f = Foo.new
+        [f.send(methods[1]), f.send(methods[2]), f.send(methods[3])]
+        "#,
+        );
+        run_test_with_prelude(
+            r#"
+            o.send(:f){ 3 }
+            "#,
+            r#"
+            class C
+                def f
+                    yield
+                end
+            end
+            o = C.new
+            "#,
+        );
+    }
+
+    #[test]
+    fn object_send() {
+        run_test_with_prelude(
+            r##"
+        o = C.new
+        [o.send(:foo), o.send("foo"), o.send(:bar, 20), o.send(*[:foo]), o.send(*["bar", 100])]
+        "##,
+            r##"
+        class C
+            def foo
+                1
+            end
+            def bar(x)
+                x
+            end
+        end
+        "##,
+        );
+        run_test_error(
+            r##"
+        class C
+            def foo
+                1
+            end
+        end
+        C.new.send
+        "##,
+        );
+        run_test_error(
+            r##"
+        class C
+            def foo
+                1
+            end
+        end
+        C.new.send(200, 100)
+        "##,
+        );
+        run_test_error(
+            r##"
+        class C
+            def foo
+                1
+            end
+        end
+        C.new.send(:foo, 100)
+        "##,
+        );
+        run_test_error(
+            r##"
+        class C
+            def foo
+                1
+            end
+        end
+        C.new.send(*[])
+        "##,
+        );
+        run_test_error(
+            r##"
+          class C
+            def b(x, y, z)
+              puts "x=#{[x, y, z]}"
+            end
+          end
+
+          o = C.new
+          30.times do |x|
+            if x == 28 then
+              eval('class C; def b(x); puts "x=#{x}" ;end; end')
+            end
+            o.send(:b, x, 2, 3)
+          end
+            "##,
+        );
+    }
+
+    #[test]
+    fn public_send() {
+        run_test(
+            r##"
+        class C
+          def foo; "foo"; end
+        end
+        C.new.public_send(:foo)
+        "##,
+        );
+        run_test(
+            r##"
+        class C
+          def bar(x); x * 2; end
+        end
+        C.new.public_send(:bar, 21)
+        "##,
+        );
+        run_test(
+            r##"
+        class C
+          def baz; yield + 1; end
+        end
+        C.new.public_send(:baz) { 41 }
+        "##,
+        );
+    }
+
+    #[test]
+    fn methods() {
+        run_test_with_prelude(
+            r##"
+        [
+          obj.singleton_methods(false).sort,
+          (obj.singleton_methods(true) - obj2.singleton_methods(true)).sort,
+          obj.methods(false).sort,
+          (obj.methods(true) - obj2.methods(true)).sort,
+          Foo.singleton_methods(false).sort,
+          (Foo.singleton_methods(true) - Foo2.singleton_methods(true)).sort
+        ]
+        "##,
+            r##"
+        Parent = Class.new
+
+        class <<Parent
+          private;   def private_class_parent() end
+          protected; def protected_class_parent() end
+          public;    def public_class_parent() end
+        end
+
+        Foo = Class.new(Parent)
+        Foo2 = Class.new()
+
+        class <<Foo
+          private;   def private_class_foo() end
+          protected; def protected_class_foo() end
+          public;    def public_class_foo() end
+        end
+
+        module Bar
+          private;   def private_bar()   end
+          protected; def protected_bar() end
+          public;    def public_bar()    end
+        end
+
+        obj = Foo.new
+        obj2 = Foo.new
+
+        class << obj
+          include Bar
+          private;   def private_self()   end
+          protected; def protected_self() end
+          public;    def public_self()    end
+        end
+        "##,
+        );
+    }
+
+    #[test]
+    fn define_singleton_method() {
+        // with block
+        run_test(
+            r#"
+            obj = Object.new
+            obj.define_singleton_method(:greet) { "hello" }
+            obj.greet
+            "#,
+        );
+        // with block taking arguments
+        run_test(
+            r#"
+            obj = Object.new
+            obj.define_singleton_method(:add) { |a, b| a + b }
+            obj.add(3, 4)
+            "#,
+        );
+        // returns symbol
+        run_test(
+            r#"
+            obj = Object.new
+            obj.define_singleton_method(:foo) { 42 }
+            "#,
+        );
+        // with string name
+        run_test(
+            r#"
+            obj = Object.new
+            obj.define_singleton_method("bar") { "baz" }
+            obj.bar
+            "#,
+        );
+        // with Proc
+        run_test(
+            r#"
+            obj = Object.new
+            pr = Proc.new { |x| x * 2 }
+            obj.define_singleton_method(:double, pr)
+            obj.double(5)
+            "#,
+        );
+        // with Method
+        run_test(
+            r#"
+            class Foo
+              def hello
+                "hello from Foo"
+              end
+            end
+            obj = Foo.new
+            obj.define_singleton_method(:greet, obj.method(:hello))
+            obj.greet
+            "#,
+        );
+        // with UnboundMethod
+        run_test(
+            r#"
+            class Foo
+              def hi
+                "hi"
+              end
+            end
+            obj = Foo.new
+            obj.define_singleton_method(:greet, Foo.instance_method(:hi))
+            obj.greet
+            "#,
+        );
+        // does not affect other instances
+        run_test(
+            r#"
+            a = Object.new
+            b = Object.new
+            a.define_singleton_method(:only_a) { "only a" }
+            [a.only_a, b.respond_to?(:only_a)]
+            "#,
+        );
+        // error: no block and no method given
+        run_test_error(
+            r#"
+            obj = Object.new
+            obj.define_singleton_method(:foo)
+            "#,
+        );
+        // error: wrong argument type
+        run_test_error(
+            r#"
+            obj = Object.new
+            obj.define_singleton_method(:foo, 42)
+            "#,
+        );
+    }
+
+    #[test]
+    fn initialize_copy() {
+        // initialize_copy with same object returns self
+        run_test(
+            r#"
+            class Foo
+              def initialize
+                @a = 1
+              end
+            end
+            obj = Foo.new
+            obj.send(:initialize_copy, obj).equal?(obj)
+            "#,
+        );
+        // initialize_copy with same class succeeds
+        run_test(
+            r#"
+            class Foo
+              def initialize
+                @a = 1
+              end
+            end
+            a = Foo.new
+            b = Foo.new
+            b.send(:initialize_copy, a).equal?(b)
+            "#,
+        );
+        // initialize_copy with different class raises TypeError
+        run_test_error(
+            r#"
+            class Foo; end
+            class Bar; end
+            a = Foo.new
+            b = Bar.new
+            a.send(:initialize_copy, b)
+            "#,
+        );
+    }
+
+    #[test]
+    fn initialize_clone() {
+        // initialize_clone delegates to initialize_copy
+        run_test(
+            r#"
+            class Foo
+              def initialize
+                @a = 1
+              end
+            end
+            a = Foo.new
+            b = Foo.new
+            b.send(:initialize_clone, a).equal?(b)
+            "#,
+        );
+        // initialize_clone with different class raises TypeError via initialize_copy
+        run_test_error(
+            r#"
+            class Foo; end
+            class Bar; end
+            a = Foo.new
+            b = Bar.new
+            a.send(:initialize_clone, b)
+            "#,
+        );
+        // overridden initialize_clone is called
+        run_test(
+            r#"
+            $called = []
+            class Foo
+              def initialize_clone(orig)
+                $called << :initialize_clone
+                super
+              end
+              def initialize_copy(orig)
+                $called << :initialize_copy
+                super
+              end
+            end
+            a = Foo.new
+            b = Foo.new
+            b.send(:initialize_clone, a)
+            $called
+            "#,
+        );
+    }
+
+    #[test]
+    fn initialize_dup() {
+        // initialize_dup delegates to initialize_copy
+        run_test(
+            r#"
+            class Foo
+              def initialize
+                @a = 1
+              end
+            end
+            a = Foo.new
+            b = Foo.new
+            b.send(:initialize_dup, a).equal?(b)
+            "#,
+        );
+        // initialize_dup with different class raises TypeError via initialize_copy
+        run_test_error(
+            r#"
+            class Foo; end
+            class Bar; end
+            a = Foo.new
+            b = Bar.new
+            a.send(:initialize_dup, b)
+            "#,
+        );
+        // overridden initialize_dup is called
+        run_test(
+            r#"
+            $called = []
+            class Foo
+              def initialize_dup(orig)
+                $called << :initialize_dup
+                super
+              end
+              def initialize_copy(orig)
+                $called << :initialize_copy
+                super
+              end
+            end
+            a = Foo.new
+            b = Foo.new
+            b.send(:initialize_dup, a)
+            $called
+            "#,
         );
     }
 }
